warning

+++ TheRegularEngineV2/TheRegularEngine/Graphics/DebugRenderer.cpp
@@ -76,7 +76,7 @@ namespace TRE
 
 		std::vector<int> DebugAABBIndices = { 0, 1, 2, 3, 0, 4, 5, 6, 7, 4, 0, 4, 1, 5, 2, 6, 3, 7, 0, 4 };
 
-		int VertexCount = DebugAABBVertices.size();
+		int VertexCount = (int)DebugAABBVertices.size();
 		uint32_t vertexSize = sizeof(DebugAABBVertices[0]);
 		Buffer stagingBuffer(vertexSize, VertexCount, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT);
 
@@ -91,7 +91,7 @@ namespace TRE
 		CopyBuffer(stagingBuffer.GetBuffer(), m_DebugAABBVertexBuffer->GetBuffer(), bufferSize);
 
 		//Index
-		m_AABBIndexCount = DebugAABBIndices.size();
+		m_AABBIndexCount = (uint32_t)DebugAABBIndices.size();
 
 		uint32_t indexSize = sizeof(int);
 		Buffer stagingBufferindex(indexSize, m_AABBIndexCount, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT);
@@ -118,7 +118,7 @@ namespace TRE
 		}
 		DebugSphereIndices.push_back(0); //Strip back to the first point
 
-		int VertexCount = DebugSphereVert.size();
+		int VertexCount = (int)DebugSphereVert.size();
 
 		uint32_t vertexSize = sizeof(DebugSphereVert[0]);
 		Buffer stagingBuffer(vertexSize, VertexCount, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT);
@@ -134,7 +134,7 @@ namespace TRE
 		CopyBuffer(stagingBuffer.GetBuffer(), m_DebugSphereVertexBuffer->GetBuffer(), bufferSize);
 
 		//Index
-		m_SphereIndexCount = DebugSphereIndices.size();
+		m_SphereIndexCount = (uint32_t)DebugSphereIndices.size();
 
 		uint32_t indexSize = sizeof(int);
 		Buffer stagingBufferindex(indexSize, m_SphereIndexCount, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT);
